remove user

+++ remove.ps1
@@ -17,23 +17,41 @@ function Import-ModuleIfNeeded {
 }
 
 # Function to remove a user from Azure AD based on email address
+# Function to remove a user from Azure AD based on email address
 function Remove-AzureADUser {
     param(
         [string]$EmailAddress
     )
 
-    $existingUser = Get-AzureADUser -Filter "Mail eq '$EmailAddress'"
-    if ($existingUser) {
-        Remove-AzureADUser -ObjectId $existingUser.ObjectId
-        Write-Output "User with email $EmailAddress has been successfully removed from Azure AD."
+    if (-not [string]::IsNullOrWhiteSpace($EmailAddress)) {
+        try {
+            # Fetch all users
+            $allUsers = Get-AzureADUser -All $true
+
+            # Filter users based on email address
+            $existingUser = $allUsers | Where-Object { $_.Mail -eq $EmailAddress }
+
+            if ($existingUser) {
+                # Remove the user
+                Remove-AzureADUser -ObjectId $existingUser.ObjectId
+                Write-Output "User with email $EmailAddress has been successfully removed from Azure AD."
+            } else {
+                Write-Output "User with email $EmailAddress does not exist in Azure AD."
+            }
+        } catch {
+            Write-Output "Error: $_"
+        }
     } else {
-        Write-Output "User with email $EmailAddress does not exist in Azure AD."
+        Write-Output "Error: Email address is empty."
     }
 }
+
+
 
 # Check if AzureAD module is installed, if not install it
 Import-ModuleIfNeeded -ModuleName "AzureAD"
 
+Start-Sleep -Seconds 10
 # Connect to Azure AD
 Connect-AzureAD
 
@@ -41,14 +59,17 @@ Connect-AzureAD
 $scriptDirectory = Split-Path -Parent $MyInvocation.MyCommand.Path
 
 # Read CSV and process users for removal
-$usersToRemove = Import-Csv -Path (Join-Path -Path $scriptDirectory -ChildPath "users.csv")
+$usersToRemove = Import-Csv -Path (Join-Path -Path $scriptDirectory -ChildPath "remove.csv")
 
 # Loop through users to remove
 foreach ($user in $usersToRemove) {
+    Start-Sleep -Seconds 15
     $emailToRemove = $user.'Email'
     # Check if email is provided and is in a valid format
     if (-not ([string]::IsNullOrWhiteSpace($emailToRemove)) -and $emailToRemove -match '\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b') {
         # Remove user from Azure AD
+        Write-Output "Attempt $emailToRemove has been successfully removed from Azure AD."
+
         Remove-AzureADUser -EmailAddress $emailToRemove
     } else {
         Write-Output "Invalid email address provided for removal: $emailToRemove"
